feat: export experimental createFetchival
